move mock back inside test methods
We need to construct a new instance of MockFor for every test;
otherwise it remembers which methods we stubbed and the tests break.

It would be fine to leave it in the class scope as long as it is
not static. If you declare it static, things break.

Move it into the test bodies just to make it clear that we're
always constructing a new instance.